Database

+++ DataBase/Comments/articles/views.py
@@ -1,10 +1,12 @@
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import redirect, render
+from django.views.decorators.http import require_http_methods, require_safe, require_POST
 
 from .forms import ArticleForm, CommentForm
 from .models import Article, Comment
 
 
+@require_safe
 def index(request):
     articles = Article.objects.all()
     context = {
@@ -13,6 +15,7 @@ def index(request):
     return render(request, 'articles/index.html', context)
 
 
+@require_safe
 def detail(request, pk):
     article = Article.objects.get(pk=pk)
     comment_form = CommentForm()
@@ -27,11 +30,14 @@ def detail(request, pk):
 
 
 @login_required
+@require_http_methods(['GET', 'POST'])
 def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user
+            article.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -42,22 +48,30 @@ def create(request):
 
 
 @login_required
+@require_POST
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
+    # 삭제을 요청하는 사용자와 게시글의 작성자가 같은지 확인
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
 
 
 @login_required
+@require_http_methods(['GET', 'POST'])
 def update(request, pk):
-    article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    article = Article.objects.get(pk=pk)    
+    # 수정을 요청하는 사용자와 게시글의 작성자가 같은지 확인
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:inde')
     context = {
         'article': article,
         'form': form,
@@ -65,6 +79,8 @@ def update(request, pk):
     return render(request, 'articles/update.html', context)
 
 
+@login_required
+@require_POST
 def comments_create(request, article_pk):
     # 어떤 게시글에 작성되는 댓글인지 알려면 게시글 먼저 조회
     article = Article.objects.get(pk=article_pk)
@@ -77,6 +93,7 @@ def comments_create(request, article_pk):
         # 댓글 인스턴스는 생성하지만, DB에 저장 요청은 보내지 않고 대기
         comment = comment_form.save(commit=False)
         comment.article = article
+        comment.user = request.user
         comment.save()
         return redirect('articles:detail', article.pk)
     context = {
@@ -85,11 +102,14 @@ def comments_create(request, article_pk):
     return render(request, 'articles/deatil.html', context)
 
 
+@login_required
+@require_POST
 def comments_delete(request, article_pk, comment_pk):
     # 어떤 댓글이 삭제되는 것인지 조회
     comment = Comment.objects.get(pk=comment_pk)
     # [보조적인 방법] comment 인스턴스에서 article pk 가져오기
     # article_pk = comment.article.pk
     # 댓글 삭제
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     return redirect('articles:detail', article_pk)
